regularly report network limits
