+++ core/Cargo.toml
@@ -79,6 +79,9 @@ solana-version = { path = "../version", version = "=1.10.0" }
 static_assertions = "1.1.0"
 systemstat = "0.1.10"
 
+[target."cfg(unix)".dependencies]
+sysctl = "0.4.3"
+
 [build-dependencies]
 rustc_version = "0.4"
