+++ validator/Cargo.toml
@@ -60,7 +60,6 @@ jemallocator = {package = "tikv-jemallocator", version = "0.4.1", features = ["u
 [target."cfg(unix)".dependencies]
 libc = "0.2.112"
 signal-hook = "0.3.13"
-sysctl = "0.4.3"
 
 [package.metadata.docs.rs]
 targets = ["x86_64-unknown-linux-gnu"]
